fix(grpc): enforce shared LoRA path validation
